+++ CHANGELOG.rst
@@ -1,5 +1,15 @@
+Release 1.3.0 - Symplectic Elements
+-----------------------------------
+* Added import-to-symplectic command to create OE Articles in Symplectic-Elements
+* Remove ability for self-upload of Articles
+
+Release 1.2.11 - Bug Fix
+------------------------
+* Fixed xml parsing error with ROMEO auto-complete
+
+
 Release 1.2.10 - Supplemental Materials
---------------------------------------
+---------------------------------------
 * Added Supplemental Materials to edit form and article view
 * Fixed big with deleting repeating fields